CMakeLists.txt: Fix-ups for out-of-tree builds.
CMake really likes it when you specify a minimum version. Let's do that.
Also added the ability te build and install on Linux out-of-tree.
Mac *should* work, but I can't test it.
Windows *should* work, but it has some trouble compiling on VS2017 15.9+ because
of changes MS made to their compilier. Seeing as that issue is outside the scope
and modifications made in this PR, that is an issue to address another day.