Add draft of new UI.
The idea is that all scene switches are treated the same.
Thus only one tab is needed for all scene switches.

A user shall be able to select each condition and even combine them.
For example an existing scene sequence would be translated to the condition of
'scene is active' combined with the two actions 'wait for x' and 'switch
scene'.
But for example the condition 'scene is active' could be combined with
'window with name ... is in focus' to allow for even more complex scene
switches.

Also the priority of different scene switch entries can now be handled
independent of their underlying switching methods.