Do not block UI with wait action
To achive this the matching of macros and execution of macros was split
in two different functions.
The latter being executed while the main mutex is not being held by the
switcher thread, meaning wait actions will not block the settings UI.
This also allows use of a CV to abort a wait early (Exit of OBS /
stopping the plugin).