Disabled bogus GCC string overflow error
GCC has some logic to detect string overflow. Unfortunately, it's broken and gives many false positives.

* https://gcc.gnu.org/bugzilla/show_bug.cgi?id=114379
* https://aur.archlinux.org/packages/obs-advanced-scene-switcher

This commit just disables the compiler warning/error.

I'm not even sure if this is the right place to disable this warning, as I'm not familiar with CMake. (Will it also affect other compilers? Would other compilers complain about this flag?) But at least it manages to fully compile the project.